Modified claims inspect and par inspect to use the inspect command Moved inspect.rs to wash-lib (as claims could not import the code from the original source) Added inspect to workspace and dev-dependencies Removed boilerplate code
