+++ Cargo.toml
@@ -101,6 +101,7 @@ futures = "0.3"
 heck = "0.4"
 ignore = "0.4"
 indicatif = "0.17.0"
+inspect = "0.1.0"
 log = "0.4"
 nkeys = "0.2.0"
 oci-distribution = { version = "0.9.1", default-features = false }
