+++ src/main.rs
@@ -10,7 +10,6 @@ use ctl::CtlCliCommand;
 use ctx::CtxCommand;
 use down::DownCommand;
 use generate::NewCliCommand;
-use inspect::InspectCliCommand;
 use keys::KeysCliCommand;
 use par::ParCliCommand;
 use reg::RegCliCommand;
@@ -18,6 +17,7 @@ use serde_json::json;
 use smithy::{GenerateCli, LintCli, ValidateCli};
 use up::UpCommand;
 use wash_lib::cli::claims::ClaimsCliCommand;
+use wash_lib::cli::inspect::InspectCliCommand;
 use wash_lib::cli::{CommandOutput, OutputKind};
 use wash_lib::drain::Drain as DrainSelection;
 
@@ -32,7 +32,6 @@ mod ctx;
 mod down;
 mod drain;
 mod generate;
-mod inspect;
 mod keys;
 mod par;
 mod reg;
@@ -151,7 +150,9 @@ async fn main() {
         CliCommand::Down(down_cli) => down::handle_command(down_cli, output_kind).await,
         CliCommand::Drain(drain_cli) => drain::handle_command(drain_cli),
         CliCommand::Gen(generate_cli) => smithy::handle_gen_command(generate_cli),
-        CliCommand::Inspect(inspect_cli) => inspect::handle_command(inspect_cli, output_kind).await,
+        CliCommand::Inspect(inspect_cli) => {
+            wash_lib::cli::inspect::handle_command(inspect_cli, output_kind).await
+        }
         CliCommand::Keys(keys_cli) => keys::handle_command(keys_cli),
         CliCommand::Lint(lint_cli) => smithy::handle_lint_command(lint_cli).await,
         CliCommand::New(new_cli) => generate::handle_command(new_cli).await,
